+++ src/WorkflowValidation/VerificationStep.cs
@@ -12,6 +12,11 @@ namespace WorkflowValidation
         /// <param name="context"></param>
         public override void Run(WorkflowContext context)
         {
+            if (!string.IsNullOrEmpty(Name))
+            {
+                context.Log($"-> Step: {Name}");
+            }
+
             Workflow.Context = context;
             Workflow.Run();
         }