+++ src/WorkflowValidation/WorkflowBuilder.cs
@@ -5,19 +5,20 @@ namespace WorkflowValidation
     /// <summary>
     /// Builder to help create and configure a workflow
     /// </summary>
-    public class WorkflowBuilder
+    public class WorkflowBuilder : IWorkflowBuilder
     {
+        private readonly IWorkflowStep _workflow = new Workflow();
+
         /// <summary>
         /// Start the workflow with the given step
         /// </summary>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflowStep StartWith(Action step)
+        public IWorkflowStep StartWith(Action step)
         {
-            var workflow = new Workflow();
-            workflow.SetStep(new Step(step));
+            _workflow.SetStep(new Step(step));
 
-            return workflow;
+            return _workflow;
         }
 
         /// <summary>
@@ -26,14 +27,13 @@ namespace WorkflowValidation
         /// <param name="description"></param>
         /// <param name="work"></param>
         /// <returns></returns>
-        public static IWorkflowStep StartWith(string description, Action work)
+        public IWorkflowStep StartWith(string description, Action work)
         {
-            var workflow = new Workflow();
-            workflow.SetStep(new Step(work)
+            _workflow.SetStep(new Step(work)
                 .SetName(description)
             );
 
-            return workflow;
+            return _workflow;
         }
 
         /// <summary>
@@ -41,12 +41,11 @@ namespace WorkflowValidation
         /// </summary>
         /// <param name="step"></param>
         /// <returns></returns>
-        public static IWorkflowStep StartWith(Action<WorkflowContext> step)
+        public IWorkflowStep StartWith(Action<WorkflowContext> step)
         {
-            var workflow = new Workflow();
-            workflow.SetStep(new Step(step));
+            _workflow.SetStep(new Step(step));
 
-            return workflow;
+            return _workflow;
         }
 
         /// <summary>
@@ -55,14 +54,28 @@ namespace WorkflowValidation
         /// <param name="description"></param>
         /// <param name="work"></param>
         /// <returns></returns>
-        public static IWorkflowStep StartWith(string description, Action<WorkflowContext> work)
+        public IWorkflowStep StartWith(string description, Action<WorkflowContext> work)
         {
-            var workflow = new Workflow();
-            workflow.SetStep(new Step(work)
+            _workflow.SetStep(new Step(work)
                 .SetName(description)
             );
 
-            return workflow;
+            return _workflow;
+        }
+
+        /// <summary>
+        /// Setup for a new workflow
+        /// </summary>
+        /// <param name="setup"></param>
+        /// <returns></returns>
+        public IWorkflowBuilder SetupWorkflow(Action<IWorkflowSetupBuilder> setup)
+        {
+            var description = new WorkflowSetup();
+            setup(description);
+
+            _workflow.WorkflowSetup = description;
+
+            return this;
         }
     }
 }