+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
+## vNext
+### Added
+- Setup for workflows containing a description that gets output
+  
+### Changed
+- Breaking: Refactor Static StartWith methods from WorkflowBuilder to Workflow
+ 
 ## v1.0.1
 ### Added
 - Add multiple Assertions to Verificationstep